final changes

+++ FaceyPhotos/Data/Sale.cs
@@ -16,7 +16,7 @@ namespace FaceyPhotos.Data
         public int ProductId { get; set; }
 
         public int Quantity { get; set; }
-        public double price { get; set; }
+        public float price { get; set; }
